Move IPolls utility to the utility module
