+++ src/collective/polls/content/poll.py
@@ -5,7 +5,7 @@ from collective.polls.config import COOKIE_KEY
 from collective.polls.config import MEMBERS_ANNO_KEY
 from collective.polls.config import PERMISSION_VOTE
 from collective.polls.config import VOTE_ANNO_KEY
-from collective.polls.polls import IPolls
+from collective.polls.utility import IPolls
 from collective.polls.widgets.enhancedtextlines import EnhancedTextLinesFieldWidget
 from plone import api
 from plone.autoform import directives as form
